feat: seed the canvas at startup only, without stirring it afterwards
The idle feed kept injecting dye whenever the pointer went quiet, so the
canvas never settled — drops appeared on their own indefinitely. Drop it: the
startup burst alone answers the black-canvas problem, and what happens after
belongs to the person at the pointer.

`ambient.ts` loses its second source and becomes `seed.ts`; the `?ambient=`
three-way switch collapses back to `?seed=off`, which the e2e drag case still
needs to attribute dye to the drag.
