+++ src/FluidCanvas.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useRef, useState } from "react";
 
-import { ambientSources, IdleSplatter, seedSplats } from "@/fluid/ambient";
 import { MAX_STEPS_PER_FRAME, TIME_STEP } from "@/fluid/config";
 import { GpuUnavailableError, initGpu } from "@/fluid/gpu";
 import { PointerTracker } from "@/fluid/pointer";
 import { createFluidRenderer } from "@/fluid/renderer";
+import { seedEnabled, seedSplats } from "@/fluid/seed";
 import type { FluidRenderer, Splat } from "@/fluid/types";
 
 /** Cap the backing store on high-DPI displays: past 2x the extra pixels cost
@@ -87,11 +87,11 @@ export const FluidCanvas = () => {
 
       let previous = performance.now();
       let owed = 0;
-      const ambient = ambientSources(window.location.search);
-      const idle = new IdleSplatter();
       // Held until the first step runs: the loop may render before enough time
       // has accumulated for a step, and the burst has to land inside one.
-      let pending: Splat[] = ambient.seed ? seedSplats(Math.random) : [];
+      let pending: Splat[] = seedEnabled(window.location.search)
+        ? seedSplats(Math.random)
+        : [];
 
       const loop = (now: number): void => {
         owed += (now - previous) / 1000;
@@ -104,17 +104,9 @@ export const FluidCanvas = () => {
         owed -= steps * TIME_STEP;
 
         const dragged = pointer.consume();
-        if (dragged !== null) {
-          pending.push(dragged);
-          idle.notifyActivity();
-        }
+        if (dragged !== null) pending.push(dragged);
 
         for (let step = 0; step < steps; step++) {
-          const drifted = ambient.idle
-            ? idle.step(TIME_STEP, Math.random)
-            : null;
-          if (drifted !== null) pending.push(drifted);
-
           // Drained into the first step that runs: replaying the frame's splats
           // on each catch-up step would multiply their force by however far
           // behind the loop had fallen.
